Finally (hopefully) fixed issue with persistence time out. Needed to add hibernate-c3p0
Switched from floats to doubles for lat/long (more decimal places)

Added XML loader service (injected). User copies xml files to server
folder (../geoshare_xml). Rest call then kicks off the upload.

Password needed (which is visible in source code! but the files are
deleted straight after so it wont do anything).

Moved JRE up classpath due to weird issue with XML.

+++ src/main/java/com/geoshare/entities/Note.java
@@ -1,51 +1,64 @@
 package com.geoshare.entities;
 
-import javax.persistence.*;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
 
 @Entity
-@Table(name="Note")
-public class Note {
-	
-	   @Id
-	   @GeneratedValue(strategy = GenerationType.AUTO) 		   
-	   private int id;
-	   
-	   private String text; 
-	   
-	   private float latitude;
-	   
-	   private float longitude;
-	
-	   public int getId( ) {
-	      return id;
-	   }
-	   
-	   public void setId(int id) {
-	      this.id = id;
-	   }
-	   
-	   public String getText( ) {
-	      return text;
-	   }
-	   
-	   public void setText(String text) {
-	      this.text = text;
-	   }
-	   
-	   public float getLatitude() {
-		   return latitude;
-	   }
-	
-		public void setLatitude(float latitude) {
-			this.latitude = latitude;
-		}
-	
-		public float getLongitude() {
-			return longitude;
-		}
-	
-		public void setLongitude(float longitude) {
-			this.longitude = longitude;
-		}
-	
+@Table(name = "Note")
+public class Note
+{
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.AUTO)
+	private int id;
+
+	private String text;
+
+	private double latitude;
+
+	private double longitude;
+
+	public int getId()
+	{
+		return id;
+	}
+
+	public void setId(int id)
+	{
+		this.id = id;
+	}
+
+	public String getText()
+	{
+		return text;
+	}
+
+	public void setText(String text)
+	{
+		this.text = text;
+	}
+
+	public double getLatitude()
+	{
+		return latitude;
+	}
+
+	public void setLatitude(double latitude)
+	{
+		this.latitude = latitude;
+	}
+
+	public double getLongitude()
+	{
+		return longitude;
+	}
+
+	public void setLongitude(double longitude)
+	{
+		this.longitude = longitude;
+	}
+
 }
